Runs arbitrary JavaScript
If an application cannot be identified from the executed command then it
will be processed as JavaScript. If this returns a non-undefined value
which is different to the input then this value is considered to be the
output of the command. Otherwise it will display as a 'command not
recognised' error.